#1376: gossiplb: move enums to a separate file
